UDP client moved to main program, starting threads

+++ config.hpp
@@ -7,7 +7,7 @@ static int UDP_PORT = 3382; //default 3382, change with -u
 static int INTERFACE_PORT = 3637; //default 3637, change with -U
 static int FIND_INTERVAL = 1; //default 1, change with -t
 static int SERVICES_INTERVAL = 10; //default 10, change with -T
-static int REFRESH_TIME = 1; //default 1, change with -v
+static float REFRESH_TIME = 1; //default 1, change with -v
 static bool DNS_SD = false; //default faulse, change with -s
 
 #endif
